+++ internal/asm/amd64/impl.go
@@ -1,4 +1,4 @@
-package amd64
+package asm_amd64
 
 import (
 	"bytes"
