+++ internal/asm/amd64/impl_test.go
@@ -1,4 +1,4 @@
-package amd64
+package asm_amd64
 
 import (
 	"encoding/binary"
