+++ internal/wasm/jit/arch_amd64.go
@@ -1,5 +1,10 @@
 package jit
 
+import (
+	wasm "github.com/tetratelabs/wazero/internal/wasm"
+	"github.com/tetratelabs/wazero/internal/wazeroir"
+)
+
 // init initializes variables for the amd64 architecture
 func init() {
 	newArchContext = newArchContextImpl
@@ -11,3 +16,14 @@ type archContext struct{}
 
 // newArchContextImpl implements newArchContext for amd64 architecture.
 func newArchContextImpl() (ret archContext) { return }
+
+func init() {
+	unreservedGeneralPurposeIntRegisters = amd64UnreservedGeneralPurposeIntRegisters
+	unreservedGeneralPurposeFloatRegisters = amd64UnreservedGeneralPurposeFloatRegisters
+}
+
+// newCompiler returns a new compiler interface which can be used to compile the given function instance.
+// Note: ir param can be nil for host functions.
+func newCompiler(f *wasm.FunctionInstance, ir *wazeroir.CompilationResult) (compiler, error) {
+	return newAmd64Compiler(f, ir)
+}
