+++ internal/wasm/jit/jit_numeric_test.go
@@ -38,7 +38,7 @@ func TestCompiler_compileConsts(t *testing.T) {
 					env := newJITEnvironment()
 
 					// Build code.
-					compiler := env.requireNewCompiler(t, nil)
+					compiler := env.requireNewCompiler(t, newCompiler, nil)
 					err := compiler.compilePreamble()
 					require.NoError(t, err)
 
@@ -121,7 +121,7 @@ func TestCompiler_compile_Add_Sub_Mul(t *testing.T) {
 						x1, x2 := values[0], values[1]
 						t.Run(fmt.Sprintf("x1=0x%x,x2=0x%x", x1, x2), func(t *testing.T) {
 							env := newJITEnvironment()
-							compiler := env.requireNewCompiler(t, nil)
+							compiler := env.requireNewCompiler(t, newCompiler, nil)
 							err := compiler.compilePreamble()
 							require.NoError(t, err)
 
@@ -281,7 +281,7 @@ func TestCompiler_compile_And_Or_Xor_Shl_Rotl_Rotr(t *testing.T) {
 						x1, x2 := values[0], values[1]
 						t.Run(fmt.Sprintf("x1=0x%x,x2=0x%x", x1, x2), func(t *testing.T) {
 							env := newJITEnvironment()
-							compiler := env.requireNewCompiler(t, nil)
+							compiler := env.requireNewCompiler(t, newCompiler, nil)
 							err := compiler.compilePreamble()
 							require.NoError(t, err)
 
@@ -406,7 +406,7 @@ func TestCompiler_compileShr(t *testing.T) {
 					x1, x2 := values[0], values[1]
 					t.Run(fmt.Sprintf("x1=0x%x,x2=0x%x", x1, x2), func(t *testing.T) {
 						env := newJITEnvironment()
-						compiler := env.requireNewCompiler(t, nil)
+						compiler := env.requireNewCompiler(t, newCompiler, nil)
 						err := compiler.compilePreamble()
 						require.NoError(t, err)
 
@@ -524,7 +524,7 @@ func TestCompiler_compile_Le_Lt_Gt_Ge_Eq_Eqz_Ne(t *testing.T) {
 						}
 						t.Run(fmt.Sprintf("x1=0x%x,x2=0x%x", x1, x2), func(t *testing.T) {
 							env := newJITEnvironment()
-							compiler := env.requireNewCompiler(t, nil)
+							compiler := env.requireNewCompiler(t, newCompiler, nil)
 							err := compiler.compilePreamble()
 							require.NoError(t, err)
 
@@ -736,7 +736,7 @@ func TestCompiler_compile_Clz_Ctz_Popcnt(t *testing.T) {
 						}
 						t.Run(name, func(t *testing.T) {
 							env := newJITEnvironment()
-							compiler := env.requireNewCompiler(t, nil)
+							compiler := env.requireNewCompiler(t, newCompiler, nil)
 							err := compiler.compilePreamble()
 							require.NoError(t, err)
 
@@ -927,7 +927,7 @@ func TestCompiler_compile_Min_Max_Copysign(t *testing.T) {
 				x1, x2 := vs[0], vs[1]
 				t.Run(fmt.Sprintf("x1=%f_x2=%f", x1, x2), func(t *testing.T) {
 					env := newJITEnvironment()
-					compiler := env.requireNewCompiler(t, nil)
+					compiler := env.requireNewCompiler(t, newCompiler, nil)
 					err := compiler.compilePreamble()
 					require.NoError(t, err)
 
@@ -1233,7 +1233,7 @@ func TestCompiler_compile_Abs_Neg_Ceil_Floor_Trunc_Nearest_Sqrt(t *testing.T) {
 				v := v
 				t.Run(fmt.Sprintf("%f", v), func(t *testing.T) {
 					env := newJITEnvironment()
-					compiler := env.requireNewCompiler(t, nil)
+					compiler := env.requireNewCompiler(t, newCompiler, nil)
 					err := compiler.compilePreamble()
 					require.NoError(t, err)
 
@@ -1333,7 +1333,7 @@ func TestCompiler_compile_Div_Rem(t *testing.T) {
 						x1, x2 := values[0], values[1]
 						t.Run(fmt.Sprintf("x1=0x%x,x2=0x%x", x1, x2), func(t *testing.T) {
 							env := newJITEnvironment()
-							compiler := env.requireNewCompiler(t, nil)
+							compiler := env.requireNewCompiler(t, newCompiler, nil)
 							err := compiler.compilePreamble()
 							require.NoError(t, err)
