+++ internal/wasm/jit/jit_test.go
@@ -137,13 +137,16 @@ func (j *jitEnv) exec(code []byte) {
 	)
 }
 
-func (j *jitEnv) requireNewCompiler(t *testing.T, functype *wasm.FunctionType) compilerImpl {
+// newTestCompiler allows us to test a different architecture than the current one.
+type newTestCompiler func(f *wasm.FunctionInstance, ir *wazeroir.CompilationResult) (compiler, error)
+
+func (j *jitEnv) requireNewCompiler(t *testing.T, fn newTestCompiler, functype *wasm.FunctionType) compilerImpl {
 	// golang-asm is not goroutine-safe so we take lock until we complete the compilation.
 	// TODO: delete after https://github.com/tetratelabs/wazero/issues/233
 	assemblerMutex.Lock()
 
 	requireSupportedOSArch(t)
-	c, err := newCompiler(
+	c, err := fn(
 		&wasm.FunctionInstance{Module: j.moduleInstance, Kind: wasm.FunctionKindWasm, Type: functype},
 		&wazeroir.CompilationResult{LabelCallers: map[string]uint32{}},
 	)
